refactor: remove v1 useUserInfo, server routers
