+++ apps/frontend/src/pages/commands/sync.vue
@@ -8,8 +8,8 @@ import { useSyncMetadata } from '../../apis/commands/useSyncMetadata'
 import NeedLogin from '../../components/NeedLogin.vue'
 import ChatSelector from '../../components/sync/ChatSelector.vue'
 import SyncStatus from '../../components/sync/SyncStatus.vue'
-import { useSessionStore } from '../../composables/v2/useSessionV2'
 import { useChats } from '../../store/useChats'
+import { useSessionStore } from '../../store/useSessionV2'
 
 const { t } = useI18n()
 const chatStore = useChats()
@@ -18,7 +18,7 @@ const { loadChats } = chatStore
 const { executeChatsSync, currentCommand: chatsSyncCommand, syncProgress: chatsSyncProgress } = useSyncChats()
 const { executeMetadataSync, currentCommand: metadataSyncCommand, syncProgress: metadataSyncProgress } = useSyncMetadata()
 const sessionStore = useSessionStore()
-const { isConnected } = storeToRefs(sessionStore)
+const { isLoggedIn } = storeToRefs(sessionStore)
 
 const selectedChats = ref<number[]>([])
 const priorities = ref<Record<number, number>>({})
@@ -34,7 +34,7 @@ function getChatTitle(chatId: number) {
 }
 
 async function startSync() {
-  if (!isConnected.value) {
+  if (!isLoggedIn.value) {
     toast.error(t('component.sync_command.not_connect'))
     return
   }
@@ -70,7 +70,7 @@ async function confirmPriorities() {
 }
 
 async function syncMetadata() {
-  if (!isConnected.value) {
+  if (!isLoggedIn.value) {
     toast.error(t('component.sync_command.not_connect'))
     return
   }
@@ -110,14 +110,14 @@ watch(() => currentCommand.value?.status, (status) => {
 // Lifecycle
 onMounted(async () => {
   await loadChats()
-  if (!isConnected.value)
+  if (!isLoggedIn.value)
     showConnectButton.value = true
 })
 </script>
 
 <template>
   <div class="space-y-4">
-    <NeedLogin :is-connected="isConnected" />
+    <NeedLogin :is-connected="isLoggedIn" />
 
     <div class="flex items-center justify-between">
       <h3 class="text-lg font-medium">
@@ -129,14 +129,14 @@ onMounted(async () => {
         </span>
         <button
           class="rounded-md bg-blue-500 px-4 py-2 text-white disabled:cursor-not-allowed hover:bg-blue-600 disabled:opacity-50"
-          :disabled="!isConnected"
+          :disabled="!isLoggedIn"
           @click="syncMetadata"
         >
           {{ t('component.sync_command.metadata_sync') }}
         </button>
         <button
           class="rounded-md bg-blue-500 px-4 py-2 text-white disabled:cursor-not-allowed hover:bg-blue-600 disabled:opacity-50"
-          :disabled="selectedChats.length === 0 || !isConnected"
+          :disabled="selectedChats.length === 0 || !isLoggedIn"
           @click="startSync"
         >
           {{ t('component.sync_command.start_sync') }}
