+++ apps/server/src/index.ts
@@ -20,8 +20,6 @@ import { setupCommandRoutes } from './routes/commands'
 import { setupConfigRoutes } from './routes/config'
 import { setupMessageRoutes } from './routes/message'
 import { setupSearchRoutes } from './routes/search'
-import { setupUserInfoRoutes } from './routes/user-info'
-import { setupWsAuthRoutes } from './routes/ws-auth'
 import { createErrorResponse } from './utils/response'
 
 export type * from './types'
@@ -64,8 +62,6 @@ function setupRoutes(app: App) {
   setupConfigRoutes(app)
   setupMessageRoutes(app)
   setupSearchRoutes(app)
-  setupWsAuthRoutes(app)
-  setupUserInfoRoutes(app)
 }
 
 // Server configuration
