+++ apps/server/src/types/apis/index.ts
@@ -1,7 +1,5 @@
-export * from './auth'
 export * from './command'
 export * from './embed'
 export * from './export'
 export * from './search'
 export * from './sync'
-export * from './user-info'
