--- apps/server/src/types/apis/user-info.ts
@@ -1,9 +1,0 @@
-export interface UserInfo {
-  id: string
-  firstName: string
-  lastName: string
-  username: string
-  photoUrl?: string
-}
-
-export interface UserInfoResponse extends UserInfo {}
